requirements.txt
Flask==2.2.5
flask-cors==3.0.10
python-dotenv==1.0.0
firebase-admin==6.0.1
google-cloud-storage==2.11.0
requests==2.31.0
